vehicle_system/requirements.txt
# RK3568 Remote Rover System - Python Dependencies

# Core dependencies
python-version>=3.8

# Video and Image Processing
opencv-python>=4.8.0
numpy>=1.24.0

# Audio Processing
pyaudio>=0.2.13
librosa>=0.10.0
scipy>=1.10.0

# WebRTC and Real-time Communication
aiortc>=1.5.0
aiohttp>=3.8.0
av>=10.0.0

# Hardware Control
RPi.GPIO>=0.7.0  # For GPIO control (if applicable)
pyserial>=3.5  # For serial communication with motor driver

# Logging and Monitoring
python-logging-loki>=0.3.2

# Configuration Management
pyyaml>=6.0
python-dotenv>=1.0.0

# Async Framework
asyncio-contextmanager>=1.0.0
aiofiles>=23.0.0

# Performance and Optimization
numba>=0.57.0

# Testing
pytest>=7.3.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Development Tools
black>=23.0.0
flake8>=6.0.0
mypy>=1.0.0
